removed the question: "Do you want to see all questions"
This was the first questions the user got after executing one of the scripts.
I recognized that this question is obsolete because the user can of course get
the same thing with the regular expression '.*'. With this pattern there will
also be the count of the matched questions which might also be interesting.